% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMcentroid.R
\name{ENMcentroid}
\alias{ENMcentroid}
\title{Grid centroid calculation}
\usage{
ENMcentroid(
  radir,
  proname = F,
  format = "tif",
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{radir}{栅格所在文件夹的路径, 可以是多个文件夹.}

\item{proname}{逻辑值, 栅格所在文件夹的名称是否为投影时期名称. 如果为T, 则生成的数据框包含投影时期.}

\item{format}{栅格格式}

\item{outdir}{输出文件路径}

\item{parallel}{是否并行}

\item{ncpu}{并行时使用的cpu数}
}
\value{
数据框
}
\description{
计算适生区栅格的质心. 栅格应为二元栅格, 即1代表适生区, 0代表非适生区.
}
\examples{
df <- ENMcentroid(
  radir = "F:/sdm/maxent/32/1/p",
  proname = F,
  format = "tif",
  parallel = T,
  ncpu = 2
)
df

}
