% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMgbif_search.R
\name{ENMgbif_search}
\alias{ENMgbif_search}
\title{Download species data from GBIF and perform routine cleaning}
\usage{
ENMgbif_search(
  spnames,
  range = NULL,
  uncertainty = 10000,
  rWCVP = TRUE,
  dist = 1,
  outdir = NULL,
  filename
)
}
\arguments{
\item{spnames}{character string or character vector. 要下载物种的学名.}

\item{range}{character string. 用国家代码(ISO 3166-1 alpha2)或WKT多边形(逆时针)表示的地理范围. 如果为NULL，则在全球范围下载.}

\item{uncertainty}{numerical. 不确定性范围, 单位m. 当coordinateUncertaintyInMeters列已经评估时允许的最大不确定性范围.}

\item{rWCVP}{logical. 是否使用rWCVP包提供的原产地范围进行过滤.}

\item{dist}{numerical. 缓冲区范围,单位°。仅rWCVP == TRUE时生效. 位于缓冲区范围内的非原产地坐标点将被保留, 需要注意当位于不同纬度时1°代表的实际地理距离不同.}

\item{outdir}{character string. 输出文件路径.}

\item{filename}{character string. 输出文件名. 不包括后缀, 格式默认为.csv.}
}
\value{
从gbif下载的原始数据压缩包及经过清洗的csv格式的物种发生数据集.
}
\description{
使用rgbif包从GBIF下载最多500条物种发生数据并进行常规数据清洗. 主要用于物种分布建模的数据演示使用. 如果想要获得完整下载, 请使用\code{\link[TBlabENM]{ENMgbif_download}}.
}
\details{
本函数主要参考\url{https://docs.ropensci.org/rgbif/articles/getting_occurrence_da.html}的流程下载和清洗数据. 主要过程如下：

\strong{1. rgbif包下载数据}
\itemize{
\item pred_default() #删除默认的地理空间问题, 仅保留带有坐标的记录, 仅保留存在的记录, 移除化石和活体标本记录. See \code{\link[rgbif]{pred_default}}.
\item pred_or(pred_lt("coordinateUncertaintyInMeters", uncertainty), pred_isnull("coordinateUncertaintyInMeters")) #对不确定性范围进行筛选并保留为NULL的记录.
\item pred_or(pred_not(pred_in("establishmentMeans", c("MANAGED","INTRODUCED"))), pred_isnull("establishmentMeans")) #删除动植物园栽培或引进的记录并保留为NULL的记录.
}

\strong{2. CoordinateCleaner包清洗数据}
\itemize{
\item cc_cen(buffer = 5000)  # remove country centroids within 5km. See \code{\link[CoordinateCleaner]{cc_cen}}.
\item cc_cap(buffer = 5000)  # remove capitals centroids within 5km. See \code{\link[CoordinateCleaner]{cc_cap}}.
\item cc_inst(buffer = 5000)  # remove zoo and herbaria within 5km. See \code{\link[CoordinateCleaner]{cc_inst}}.
\item cc_sea() #remove from ocean. See \code{\link[CoordinateCleaner]{cc_sea}}.
\item cc_val()  #removes or flags non-numeric and not available coordinates. See \code{\link[CoordinateCleaner]{cc_val}}.
\item cc_equ()  #删除经纬度相等的坐标. See \code{\link[CoordinateCleaner]{cc_equ}}.
\item cc_zero(buffer = 0.5)  #删除经度或纬度为零及其半径0.5°内的记录. See \code{\link[CoordinateCleaner]{cc_zero}}.
}

\strong{3. dplyr包筛选数据}
\itemize{
\item filter(year > 1950) #只选择1950年后的记录.
\item filter(coordinateprecision < 0.01 | is.na(coordinateprecision)) #删除坐标精度小于0.01的点并保留为NA的点.
\item filter(!coordinateuncertaintyinmeters \%in\% c(301,3036,999,9999)) #删除已知有错误的坐标不确定度.
}

\strong{4. rWCVP包过滤非原产地的分布点}

如果rWCVP == TRUE还将使用rWCVP包过滤非原产地的记录. 更多细节参考\url{https://matildabrown.github.io/rWCVP/articles/coordinate-cleaning.html}.
}
\examples{
\donttest{ENMgbif_search(spnames = c("Phoebe formosana", "Decaisnea insignis", "Ginkgo biloba"),
                 range = "POLYGON((72.2 56, 137.5 56, 137.5 0, 72.2 0, 72.2 56))", #按给定的地理范围下载
               # range = "CN", #按国家代码下载
               # range = NULL, #全球区域下载
                 uncertainty = 10000,
                 rWCVP = TRUE,
                 dist = 1,
                 outdir = "./gbif",
                 filename = "tidy_gbif")
              }
}
\references{
\url{https://docs.ropensci.org/rgbif/articles/getting_occurrence_da.html}

\url{https://matildabrown.github.io/rWCVP/articles/coordinate-cleaning.html}
}
\seealso{
\code{\link[=ENMgbif_download]{ENMgbif_download()}}
}
