% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMPic.R
\name{ENMPic}
\alias{ENMPic}
\title{Map Suitable Areas}
\usage{
ENMPic(
  parameters,
  x,
  key,
  threshold = "T_MTSSte",
  resultdir = "./TBlabENM/result",
  wordmap = NULL,
  regionmap = NULL,
  color = NULL,
  legend = TRUE,
  format = "jpeg",
  path
)
}
\arguments{
\item{x}{数值向量。表示要进行转换的物种的序号。这个序号来自ENMspdata()生成的标准文件。}

\item{threshold}{划分适生区的阈值。可以为"T_MTSSte", "T_MTSStr"（依据spdata）
或者为一个0到1的数值型，比如0.2。}

\item{resultdir}{maxent结果文件路径。例如paste0(getwd(), "/TBlabENM/result")}

\item{wordmap}{世界地图}

\item{regionmap}{区域地图，比如研究区域的地图}

\item{color}{指定不同适生区等级的颜色。5个等级，需要指定5种颜色，分别代表适生等级从低到高。}

\item{legend}{逻辑值，是否需要图例，默认添加图例}

\item{format}{结果保存格式,(e.g. png), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{结果保存路径}

\item{spdata}{由ENMspdata()生成的标准文件。也可手动创建或修改，但是不建议。
当maxent结果文件出现变动后，强烈建议重新运行ENMspdata()。}

\item{proname}{想要绘图的时期名称。}
}
\value{
按照阈值进行划分的适生区图
}
\description{
根据阈值绘制适生区图，阈值以下代表非适生区。可以采用固定阈值，也可以自由指定。
}
\examples{
#读取.spdata.csv文件
spdata <- read.csv(paste0(system.file(package="TBlabENM"),"/extdata/.spdata.csv"))
#绘图
ENMPic(x = 1:10,
      spdata = spdata,
      proname = "present",
      threshold = 0.2,
      resultdir = "./TBlabENM/result",
      wordmap = "D:/Desktop/ENMdata/worldmap/带投影坐标系/world_pro.shp",
      regionmap = "D:/Desktop/ENMdata/Chinamap/2022年省界/sheng2022.shp",
      #color = c("", "", "", "", ""),
      legend = TRUE,
      format = "jpeg",
      path= "./TBlabENM/sppicture")
}
