% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMplot.R
\name{ENMplot}
\alias{ENMplot}
\title{Draw pictures of ENM related results}
\usage{
ENMplot(
  raster,
  rcl = NULL,
  color = NULL,
  wordmap = NULL,
  regionmap = NULL,
  label = NULL,
  legend = TRUE,
  Legend_title = "Value",
  format = "jpeg",
  filename = "myplot.jpeg",
  path
)
}
\arguments{
\item{raster}{栅格文件完整路径}

\item{rcl}{向量，用于指定重分类的区间，}

\item{color}{指定不同适生区等级的颜色。默认5个等级。}

\item{wordmap}{世界地图}

\item{regionmap}{区域地图，比如研究区域的地图}

\item{label}{标注文字，例如时期的标注}

\item{legend}{逻辑值，是否需要图例，默认添加图例}

\item{Legend_title}{图例标题}

\item{format}{结果保存格式,(e.g. png), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{filename}{保存的文件名称}

\item{path}{结果保存路径}
}
\value{
picture
}
\description{
对相关结果进行分类的单张图片绘制，比如适生区图、避难所图、丰富度图等。
}
\examples{
#绘图
ENMplot(raster = "G:/Pittosporum/TBlabENM/sr_refuges_future-past/spp-refuges.tif",
        rcl = c(0, 5, 10, 15, 20, 24),
        color = NULL,
        wordmap = "D:/Desktop/ENMdata/worldmap/带投影坐标系/world_pro.shp",
        regionmap = "D:/Desktop/ENMdata/Chinamap/2022年省界/sheng2022.shp",
        label = NULL,
        legend = TRUE,
        Legend_title = "Value",
        format = "jpeg",
        filename = "2.jpeg",
        path = "G:/Pittosporum")
}
