% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMreclass_area.R
\name{ENMreclass_area}
\alias{ENMreclass_area}
\title{Calculate the area of the reclassified raster}
\usage{
ENMreclass_area(
  reclassdir,
  crs,
  prefix = NULL,
  suffix,
  overwrite = FALSE,
  outdir = NULL
)
}
\arguments{
\item{reclassdir}{重分类栅格所在路径}

\item{crs}{给缺乏投影的栅格定义投影. 例如"epsg:4326". 栅格的投影应与环境变量保持一致, 错误的投影会得到错误的结果.}

\item{prefix}{字符型. 要重分类的栅格文件名的前缀, 用于筛选栅格.}

\item{suffix}{字符型. 要重分类的栅格文件名的后缀, 用于筛选栅格.}

\item{overwrite}{logical. If TRUE, filename is overwritten.}

\item{outdir}{输出文件夹路径.}
}
\value{
列表. reclass_area_path为计算的面积文件存储路径; area为重分类后的面积(单位KM²).
}
\description{
计算重分类栅格面积.
}
\examples{
ENMreclass_area(
reclassdir = "/ifs1/User/wuhaiyang/sdm/reclass25部分",
crs = "epsg:4326",
prefix = "wz0001",
suffix = "tif",
outdir = "/ifs1/User/wuhaiyang/sdm"
)

}
