% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMspcheck.R
\name{ENMspcheck}
\alias{ENMspcheck}
\title{Occurrence Data Check}
\usage{
ENMspcheck(
  spdir,
  evdir,
  evselectev,
  elevation,
  elrange = 500,
  deal_NA = "none",
  width,
  outdir
)
}
\arguments{
\item{spdir}{要模拟的物种文件的路径，包括文件名（.csv）。}

\item{evdir}{环境变量路径。格式为 .asc 或者 .tif。}

\item{evselectev}{字符型；用于检查数据的变量，第一个要为海拔变量。一种类型的变量选择一个即可。}

\item{elevation}{海拔范围，单位m。字符型向量。}

\item{elrange}{数值型向量。允许的海拔误差，单位m。默认500m。}

\item{deal_NA}{字符型,当发生点所在的栅格像元为缺失值时的处理方法.可选择"mv","rm","none".
当为"mv"时,对物种坐标进行微调,移动至最近的非NA值像元上，当为"rm"时,直接删除该点,当为"none"时,
不做任何处理.}

\item{width}{数值型,指定缓冲区的宽度,当deal_NA选择"mv"时,在给定宽度的缓冲区内如果栅格都为缺失值则
直接删除该点.单位通常为m.该值应大于像元的宽度，否则没有意义.}

\item{outdir}{当deal_NA选择"mv"或"rm"时,处理后的数据的保存路径.}
}
\value{
返回可能存在问题的数据，建议对这些数据进行检查。
}
\description{
检查物种发生数据是否存在缺失值以及是否在指定的海拔范围内。
}
\examples{
ENMspcheck(spdir = "./xx.csv",
           evdir = paste0(system.file(package="TBlabENM"), "/extdata/present"),
           #用于检查的变量，一种类型的变量选择一个即可
           evselectev = c("elev30", "bio01", "SU_CODE", "uvb1"),
           #海拔范围从植物志查询
           elevation = c(1000,1200),
           elrange = 500,
           deal_NA = "none")
}
\author{
吴海洋 和 田斌
}
