% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMsprichness.R
\name{ENMsprichness}
\alias{ENMsprichness}
\title{Species Richness Analysis}
\usage{
ENMsprichness(
  parameters,
  x = NULL,
  booleandir,
  key = NULL,
  bioindex = "sr",
  overwrite = FALSE,
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{parameters}{数据框，第一列为物种名(要与resultdir下的名称保持一致),其他列位置不做要求，但是必须有一列或多列
为划分适生区和非适生区的阈值。}

\item{x}{数值向量,表示物种序号,与parameters中的物种序号一致。}

\item{booleandir}{适生区和非适生区结果文件路径。}

\item{key}{要提取的关键词，一般对应投影时期}

\item{bioindex}{丰富度指数，可选sr,we,cwe}

\item{overwrite}{logical. If TRUE, filename is overwritten}

\item{outdir}{输出文件夹}

\item{parallel}{是否并行}

\item{ncpu}{并行时cpu个数}

\item{spdata}{spdata 由ENMspdata()生成的标准文件。也可手动创建或修改，但是不建议。
当maxent结果文件出现变动后，强烈建议重新运行ENMspdata()。}

\item{proname}{想要计算的物种丰富度的时期名称.}

\item{path}{输出结果路径。}
}
\value{
物种丰富度图，.tif格式。
}
\description{
根据布尔图进行叠加，计算物种丰富度。
}
\examples{
#读取.spdata.csv
spdata <- read.csv(paste0(system.file(package="TBlabENM"),"/extdata/spdata.csv"))
#丰富度计算
ENMsprichness(parameters = read.csv("F:/eblf/TBlabENM/spdata.csv"),
              x = c(3,5,7,9),
              booleandir = "F:/eblf/TBlabENM/BooleanmapT_MTSStr",
              key = list("2050RCP8.5"= "2050RCP8.5",
                         "2070RCP8.5"= "2070RCP8.5",
                         "Present"= "Present"),
              bioindex = "cwe",
              overwrite = T,
              outdir = "F:/4",
              parallel = T,
              ncpu = 2)
}
