% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMspthin.R
\name{ENMspthin}
\alias{ENMspthin}
\title{Batch spatially thin species occurence data}
\usage{
ENMspthin(spdir, thin.par, outdir = NULL)
}
\arguments{
\item{spdir}{要过滤的物种文件的路径的集合，包括文件名（.csv）.}

\item{thin.par}{数值型，地理空间过滤距离（km）.}

\item{outdir}{结果保存路径，未指定时在当前工作路径自动生成保存路径.}
}
\value{
一个文件夹,包含每个物种过滤后的csv文件和过滤前后的连接文件.
}
\description{
批量对物种发生数据进行空间过滤.
}
\details{
本函数利用\code{\link[spThin]{thin}}对物种发生数据进行地理空间过滤来降低采样偏差，并且额外
生成一个连接文件(compare开头),指明了与原始数据相比哪些记录被保留了下来.
}
\examples{
#读取物种路径列表
dir <- system.file("extdata", "species", package = "TBlabENM")
splist <- list.files(spdir, pattern = ".csv$", full.names = T)
ENMspthin(spdir = splist, thin.par = 10, outdir = NULL)
}
