% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMunityenv.R
\name{ENMunityenv}
\alias{ENMunityenv}
\title{Environment variable processing}
\usage{
ENMunityenv(
  envdir,
  proname,
  factors = NULL,
  range,
  crop = T,
  method = "bilinear",
  res = NULL,
  outformat = "tif",
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{envdir}{要处理的原始环境变量集所在路径，该路径下包括代表不同时期或地理空间的子文件夹，
每个子文件夹内存放对应的环境变量。}

\item{proname}{不同时期或地理空间的名称，与envdir下的子文件夹名称对应}

\item{factors}{分类变量名称}

\item{range}{SpatRaster, SpatVector}

\item{crop}{是否将原始变量裁剪至研究区范围}

\item{method}{连续变量重采样的方法}

\item{res}{数值型，当range为SpatVector时输出栅格的分辨率}

\item{outformat}{输出文件格式,可选asc,tif或二者组合}

\item{outdir}{输出文件夹}

\item{parallel}{是否并行}

\item{ncpu}{并行使用的cpu数}
}
\value{
返回tif或asc文件
}
\description{
将不同来源的环境变量转换为格式、分辨率、范围统一的变量，用于生态位建模。目前仅
支持tif和asc格式变量的处理。
}
\examples{
ENMunityenv(envdir = "F:/example/TBlabENM/env1",
            proname <- c("present", "lgm"),
            factors = NULL,
          # range <- rast("C:/Users/why/TBlabENM/inst/extdata/envar/tif/bio1.tif"),
            range <- vect("D:/Desktop/ENMdata/Chinamap/2022年省界/sheng2022.shp") ,
            crop = T,
            method  = "bilinear",
            res = 10000,
            outformat = "tif",
            outdir = "F:/example",
            parallel = F,
            ncpu = 2)

}
