% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_args.R
\name{maxent_args}
\alias{maxent_args}
\title{MaxEnt Models Parameters}
\usage{
maxent_args(
  replicates = 10,
  betamultiplier = 1,
  l = TRUE,
  q = TRUE,
  p = FALSE,
  t = FALSE,
  h = FALSE,
  replicatetype = "crossvalidate",
  responsecurves = TRUE,
  jackknife = TRUE,
  pictures = TRUE,
  outputgrids = FALSE,
  outputformat = "logistic"
)
}
\arguments{
\item{replicates}{模型重复次数。默认10次。}

\item{betamultiplier}{正则化乘数。默认为1。}

\item{l}{是否在模型中使用特征函数"linear"。默认为true。}

\item{q}{是否在模型中使用特征函数"quadratic"。默认为true。}

\item{p}{是否在模型中使用特征函数"product"。默认为false。}

\item{t}{是否在模型中使用特征函数"threshold"。默认为false。}

\item{h}{是否在模型中使用特征函数"hinge"。默认为false。}

\item{replicatetype}{指定重复类型。默认为"crossvalidate"。}

\item{responsecurves}{是否生成响应曲线图。默认为true。}

\item{jackknife}{是否进行折刀分析。默认为true。}

\item{pictures}{是否生成模拟图。默认为true。}

\item{outputgrids}{是否为每次模拟生成栅格结果。默认为false。}

\item{outputformat}{选择输出格式。默认为"logistic"。}
}
\value{
包含MaxEnt建模参数的字符型向量。
}
\description{
本函数生成MaxEnt模型的默认参数(TBlab)，您也可以自己指定。常与其他
建模函数连用.
}
\examples{
# 使用默认参数(TBlab)
args <- maxent_args()
args
# 自定义参数。
args <- maxent_args(
  replicates = 5,
  betamultiplier = 2
)
args
}
\author{
吴海洋
}
