% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_auto.R
\name{maxent_auto}
\alias{maxent_auto}
\title{Maxent Automation:From occurrence to maps of suitable areas}
\usage{
maxent_auto(
  spdir,
  evdir,
  myenv = NULL,
  evlist = NULL,
  factors = NULL,
  mybgfile = NULL,
  nbg = 10000,
  args = maxent_args(),
  fc,
  rm,
  r = 0.7,
  cormethod = "pearson",
  vif = T,
  vifth = 5,
  opt = NULL,
  prodir = NULL,
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{evdir}{环境变量路径,格式为.asc}

\item{myenv}{指定的一组环境变量，当提供此参数时，将不会再根据相关性进行变量筛选}

\item{evlist}{用于建模的环境变量，是环境变量路径下所有变量的子集。用下标表示，
数字代表要选用的变量，默认使用所有变量。}

\item{factors}{字符型向量。指定哪些变量是分类变量。未指定时，则默认全为连续变量。}

\item{mybgfile}{自定义的背景数据，包含两列（经度、纬度）}

\item{nbg}{随机背景点数量，当指定参数mybgfile时忽略。}

\item{args}{自定义的MaxEnt模型参数，详见\code{\link[TBlabENM]{maxent_args}}。}

\item{fc}{特征函数,取值为L,Q,P,H,T,及它们的组合}

\item{rm}{正则化乘数 数值型}

\item{r}{相关性系数}

\item{cormethod}{计算相关性的方法，取值为"pearson" (default), "kendall", or "spearman"}

\item{vif}{逻辑型，当为T时对所有候选变量组合进行方差膨胀因子检验，排除具有强共线性的组合}

\item{vifth}{方差膨胀因子阈值}

\item{opt}{选择最佳模型的指标，可选择"aicc","seq","cbi".}

\item{prodir}{用列表储存的投影文件路径，包含了投影时期的名称和与之对应的环境变量路径}

\item{outdir}{结果保存路径}

\item{parallel}{是否并行计算}

\item{ncpu}{如果并行，使用的cpu数}

\item{sp}{要模拟的一或多个物种文件的路径，包括文件名（.csv）。}
}
\value{
包含maxent相关结果的文件夹
}
\description{
本函数自动执行MaxEnt模型，当设置好参数后，首先进行相关参数的选择，包括环境变量、
正则化乘数和特征函数。接着使用选择好的最佳参数进行建模然后投影至指定的时空生成
适宜性地图。
}
\examples{
maxent_auto(sp = system.file("extdata", "species/wz0001.csv", package = "TBlabENM"),
evdir = system.file("extdata", "envar/asc", package = "TBlabENM"),
evlist = 1:7,
factors = c("dr", "fao90"),
nbg = 1000,
args = maxent_args(),
myenv = NULL,
fc = c("lph", "q", "lq"),
rm = 1:2,
r = 0.7,
vif = T,
vifth = 5,
opt = "aicc",
outdir = NULL)

}
