% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_envcb.R
\name{maxent_envcb}
\alias{maxent_envcb}
\title{Statistical Variables According to Contribution Size and Plot}
\usage{
maxent_envcb(
  parameters,
  x = NULL,
  resultdir,
  wordsize = 1,
  device = "jpeg",
  outdir = NULL
)
}
\arguments{
\item{parameters}{数据框，第一列为物种名(要与resultdir下的名称保持一致),第二列为坐标点数,第三列为使用的环境变量}

\item{x}{数值向量,表示物种序号,与parameters中的物种序号一致。}

\item{resultdir}{maxent结果文件路径。}

\item{wordsize}{用于调整图的标签字体大小，数值型。}

\item{device}{输出图片格式，or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{outdir}{输出文件路径}
}
\value{
生成"./Variable_contribution"文件夹，该文件夹包括一个按贡献大小排列变量的.csv文件
和.jpeg的最重要的三个变量贡献堆叠条形统计图。
}
\description{
本函数根据环境变量对模型的贡献大小进行排序，并统计对模型贡献最大的前三个变量的个数
绘制堆叠条形统计图。
}
\examples{
maxent_envcb(
  parameters = read.csv("F:/4/TBlabENM/maxent/parameters.csv"),
  x = 1:10,
  resultdir = "F:/4/TBlabENM/maxent",
  wordsize = 1,
  device = "jpeg",
  outdir = "F:/4"
)
}
