% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_parameter.R
\name{maxent_parameter}
\alias{maxent_parameter}
\title{Select optimum parameters for the MaxEnt models}
\usage{
maxent_parameter(
  x,
  evdir,
  myenv = NULL,
  evlist = NULL,
  factors = NULL,
  mybgfile = NULL,
  nbg = 10000,
  fc,
  rm,
  r = 0.7,
  cormethod = "pearson",
  vif = T,
  vifth = 5,
  opt = "aicc",
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{x}{要模拟的物种文件的路径，包括文件名（.csv）。}

\item{evdir}{环境变量路径,格式为.asc}

\item{myenv}{指定的一组环境变量，当提供此参数时，将不会再根据相关性进行变量筛选}

\item{evlist}{用于建模的环境变量，是环境变量路径下所有变量的子集。用下标表示，
数字代表要选用的变量，默认使用所有变量。}

\item{factors}{字符型向量。指定哪些变量是分类变量。未指定时，则默认全为连续变量。}

\item{mybgfile}{自定义的背景数据，包含两列（经度、纬度）}

\item{nbg}{随机背景点数量，当指定参数mybgfile时忽略。}

\item{fc}{特征函数,取值为L,Q,P,H,T,及它们的组合}

\item{rm}{正则化乘数 数值型}

\item{r}{相关性系数}

\item{cormethod}{计算相关性的方法，取值为"pearson" (default), "kendall", or "spearman"}

\item{vif}{逻辑型，当为T时对所有候选变量组合进行方差膨胀因子检验，排除具有强共线性的组合}

\item{vifth}{方差膨胀因子阈值}

\item{opt}{选择最佳模型的指标，可选择"aicc","seq","cbi"或NULL.当选择NULL时，则仅选择变量。}

\item{outdir}{结果保存路径}

\item{parallel}{是否并行计算}

\item{ncpu}{如果并行，使用的cpu数}
}
\value{
生成一个csv文件，包含了每种fc、rm、选择的环境变量组合的评估结果
}
\description{
本函数为MaxEnt模型选择最适合的环境变量组合、正则化乘数和特征函数组合。
}
\details{
工作流程：首先对fc和rm进行所有可能的组合，对于每个组合，使用全部变量进行建模获得变量贡献大小，
依据变量贡献大小排除与最重要变量具有强相关性的变量。然后使用保留的变量继续建模获得该组变量
贡献大小，依据变量贡献大小排除与第二重要的变量具有强相关性的变量。重复上述过程直到保留的
变量之间不具有强相关性。经过此步骤就为每个fc和rm组合获得了一组不具有强相关性且保留了贡献大
的候选变量，当vif=T时，还对这一系列候选变量组合进行方差膨胀因子分析，进一步排除具有强共线性
的组合。最后使用ENMevaluate对每个组合进行测试，获得相应的指标（例如“aicc”），根据指标选择
最佳的一组组合作为模型最终参数。
注意：当物种发生数据过少时，可能无法得出正确的AICC和CBI值，如果opt选择aicc或cbi，则使用
auc.val.avg值进行代替。
}
\examples{
maxent_parameter(x = system.file("extdata", "species/Phoebe sheareri.csv", package = "TBlabENM"),
       evdir = system.file("extdata", "envar/asc", package = "TBlabENM"),
       myenv = NULL,
       evlist = 1:7,
       factors = c("dr", "fao90"),
       nbg = 1000,
       fc = c("LPH", "Q", "LQ"),
       rm = 1:5,
       r = 0.7,
       vif = T,
       vifth = 5,
       opt = "aicc",
       parallel = F,
       ncpu = 2)
}
