% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_single.R
\name{maxent_single}
\alias{maxent_single}
\title{Perform the MaxEnt model for a single species}
\usage{
maxent_single(
  x,
  evdir,
  evlist = NULL,
  factors = NULL,
  mybgfile = NULL,
  nbg = 10000,
  args = maxent_args(),
  prodir = NULL,
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{x}{要模拟的物种文件的路径，包括文件名（.csv）。}

\item{evdir}{环境变量路径,格式为.asc}

\item{evlist}{用于建模的环境变量，是环境变量路径下所有变量的子集。用下标表示，
数字代表要选用的变量，默认使用所有变量。}

\item{factors}{字符型向量。指定哪些变量是分类变量。未指定时，则默认全为连续变量。}

\item{mybgfile}{自定义的背景点。包含两列（经度、纬度）
未指定时，随机从环境层生成。}

\item{nbg}{随机背景点数量，当指定参数mybgfile时忽略。}

\item{args}{自定义的MaxEnt模型参数，详见\code{\link[TBlabENM]{maxent_args}}。}

\item{prodir}{用列表储存的投影文件路径，包含了投影时期的名称和与之对应的环境变量路径}

\item{outdir}{输出文件路径，若未指定则在当前目录下生成./TBlabENM/maxent文件夹保存输出结果}

\item{parallel}{是否并行计算}

\item{ncpu}{如果并行，使用的cpu数}
}
\value{
包含maxent模型结果的一系列文件
}
\description{
本函数为单个物种构建MaxEnt模型.
}
\details{
如果已经对MaxEet模型参数进行了优化(例如使用\code{\link[ENMeval]{ENMevaluate}})，
用户可以使用参数args在一定程度上修改模型参数，来获得质量更高的模型。可以使用evlist
参数来选择需要的环境变量集。当提供参数prodir时，则将模型投影至指定的地理空间，否则仅拟合模型。
}
\examples{

maxent_single(
  x = system.file("extdata", "species/Phoebe sheareri.csv", package = "TBlabENM"),
  evdir = system.file("extdata", "envar/asc", package = "TBlabENM"),
  evlist = 1:7,
  factors = c("dr", "fao90"),
  nbg = 1000,
  args = maxent_args(),
  prodir = list(present = system.file("extdata", "envar/asc", package = "TBlabENM"),
              present1 = system.file("extdata", "envar/asc", package = "TBlabENM")
  ),
  parallel = T,
  ncpu =2)
}
