% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_spdata.R
\name{maxent_spdata}
\alias{maxent_spdata}
\title{Create formatted input data from the maxent results folder}
\usage{
maxent_spdata(resultdir = "./TBlabENM/result", proname = "present")
}
\arguments{
\item{resultdir}{maxent结果文件路径。}

\item{proname}{字符型向量。要提取的时期名称。根据自己的时期确定。}
}
\value{
一个格式规范的.CSV文件。
}
\description{
本函数从maxent的结果文件中提取物种名、T_MTSSte, T_MTSStr和AUC，作为进行
物种丰富度、避难所等分析的输入文件。注意：重新运行该命令会覆盖上一次的结果。
}
\examples{
#设置工作路径
setwd("D:/Desktop")
#读取参数结果
maxent_spdata(resultdir="./TBlabENM/result",
          proname = "present"
 )
}
