% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMsprichness.R
\name{ENMsprichness}
\alias{ENMsprichness}
\title{Species Richness Analysis}
\usage{
ENMsprichness(
  d,
  x = NULL,
  booleandir,
  key = NULL,
  bioindex = "sr",
  overwrite = FALSE,
  outdir = NULL,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{d}{数据框或字符向量, 提供要计算的物种名称. 为数据框时, 第一列为物种名. (物种名称要与booleandir下的表示物种存在/不存在栅格文件名保持一致)}

\item{x}{数值向量, 表示要纳入分析的物种序号, 与d中的物种序号一致. 默认使用全部物种.}

\item{booleandir}{物种的存在/不存在栅格文件所在路径. 文件名要包含d中指定的物种.}

\item{key}{列表, 用于提取booleandir中对应的栅格进行分析. 例如分别计算不同时期的物种丰富度.}

\item{bioindex}{丰富度指数, 可选sr, we, cwe中的任意一个. 如果为cwe, 同时计算sr和we.}

\item{overwrite}{logical. If TRUE, filename is overwritten.}

\item{outdir}{输出文件夹}

\item{parallel}{是否并行}

\item{ncpu}{并行时cpu个数}
}
\value{
物种丰富度图，.tif格式。
}
\description{
叠加不同物种的存在/不存在栅格，计算物种丰富度.
}
\examples{
ENMsprichness(
  d = read.csv("F:/eblf/TBlabENM/spdata.csv"),
  x = c(3, 5, 7, 9),
  booleandir = "F:/eblf/TBlabENM/BooleanmapT_MTSStr",
  key = list(
    "2050RCP8.5" = "2050RCP8.5",
    "2070RCP8.5" = "2070RCP8.5",
    "Present" = "Present"
  ),
  bioindex = "cwe",
  overwrite = T,
  outdir = "F:/4",
  parallel = T,
  ncpu = 2
)
}
