% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMunityenv.R
\name{ENMunityenv}
\alias{ENMunityenv}
\title{Environment variable processing}
\usage{
ENMunityenv(
  radir,
  ref,
  proname = NULL,
  factors = NULL,
  method = "bilinear",
  res,
  format = "tif",
  outdir = NULL,
  overwrite = F,
  parallel = F,
  ncpu = 2
)
}
\arguments{
\item{radir}{要处理的环境变量所在的路径}

\item{ref}{SpatRaster或SpatVector,作为参考研究区，如果是SpatRaster裁剪后的变量范围和分辨率与
SpatRaster一致,如果是SpatVector,裁剪后的变量范围与SpatVector一致，分辨率由res参数指定。}

\item{proname}{投影时期的名字，必须是radir下的子文件夹的名字}

\item{factors}{分类变量名，要与radir下的变量名称一致}

\item{method}{连续变量的重采样方法，对于分类变量内部设置为near，可选bilinear,cubic,near等}

\item{res}{numeric. Can be used to set the resolution of the output raster if ref is a SpatVector}

\item{format}{输出变量格式}

\item{outdir}{输出文件路径}

\item{overwrite}{logical. If TRUE, filename is overwritten}

\item{parallel}{是否并行}

\item{ncpu}{并行的cpu数}
}
\value{
栅格文件
}
\description{
将不同来源的环境变量进行格式转换、重新投影到指定的研究区.
}
\examples{
ENMunityenv(radir = "F:/example/env",
            ref = rast("F:/var2/eblf_proj/present/tif/bio1.tif"),
            proname = c("present", "acc2030ssp126","acc2030ssp245"),
            outdir = "F:/example/env",
            overwrite = T,
            parallel = T,
            ncpu = 2)

}
