% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_auto.R
\name{maxent_auto}
\alias{maxent_auto}
\title{Maxent Automation:From occurrence to maps of suitable areas}
\usage{
maxent_auto(
  spdir,
  evdir,
  myenv = NULL,
  evlist = NULL,
  factors = NULL,
  mybgfile = NULL,
  nbg = 10000,
  args = maxent_args(),
  fc,
  rm,
  r = 0.7,
  cormethod = "pearson",
  vif = T,
  vifth = 5,
  opt = "auc.val.avg",
  null_model = TRUE,
  prodir = NULL,
  outdir = NULL,
  p_ncpu = FALSE
)
}
\arguments{
\item{spdir}{要模拟的一或多个物种文件的路径，包括文件名（.csv）。}

\item{evdir}{环境变量路径,格式为.asc}

\item{myenv}{指定的一组环境变量，当提供此参数时，将不会再根据相关性进行变量筛选}

\item{evlist}{用于建模的环境变量，是环境变量路径下所有变量的子集。用下标表示，
数字代表要选用的变量，默认使用所有变量。}

\item{factors}{字符型向量。指定哪些变量是分类变量。未指定时，则默认全为连续变量。}

\item{mybgfile}{自定义的背景数据，包含两列（经度、纬度）}

\item{nbg}{随机背景点数量，当指定参数mybgfile时忽略。}

\item{args}{自定义的MaxEnt模型参数，详见\code{\link[TBlabENM]{maxent_args}}。}

\item{fc}{特征函数,取值为L,Q,P,H,T,及它们的组合}

\item{rm}{正则化乘数 数值型}

\item{r}{相关性系数}

\item{cormethod}{计算相关性的方法，取值为"pearson" (default), "kendall", or "spearman"}

\item{vif}{逻辑型，当为T时对所有候选变量组合进行方差膨胀因子检验，排除具有强共线性的组合}

\item{vifth}{方差膨胀因子阈值}

\item{opt}{选择最佳模型的指标，可选择"auc.train, cbi.train, auc.diff.avg, auc.val.avg, cbi.val.avg, or.10p.avg, or.mtp.avg, AICc"或NULL.当选择NULL时，则仅选择变量。当选择多个指标时，则按照顺序筛选最佳模型，如果所选指标具有NA，则跳过该指标，如果所有指标都含有NA，则使用auc.val.avg代替.}

\item{null_model}{逻辑值,是否进行NULL模型检验。}

\item{prodir}{用列表储存的投影文件路径，包含了投影时期的名称和与之对应的环境变量路径}

\item{outdir}{结果保存路径}

\item{p_ncpu}{是否并行计算及各部分并行计算的cpu数.共有3处可能用到并行，分别是多物种并行，
每个物种参数调优并行和每个物种不同时期投影并行.如果不使用并行，则p_ncpu = FALSE.
当使用并行时，可分别对这3处设置不同的cpu数来最大化效率.详见details.}
}
\value{
包含maxent相关结果的文件夹
}
\description{
本函数自动执行MaxEnt模型，当设置好参数后，首先进行相关参数的选择，包括环境变量、
正则化乘数和特征函数。接着使用选择好的最佳参数进行建模然后投影至指定的时空生成
适宜性地图。
}
\details{
p_ncpu参数
当电脑或服务器的cpu数较多时，可以选择并行来加快速度.例如可以设置p_ncpu = c(2,5,10)
即同时模拟2个物种，在为每个物种选择参数过程中使用5个cpu，在为每个物种执行投影时
使用10个cpu,在这个过程中最多可能使用到2*10共20个cpu,因此每处cpu设置的个数应该结合
电脑cpu数和实际模拟需要的cpu数进行合理设置.如果某个过程不想执行并行，可以设置为FALSE，
例如p_ncpu = c(FALSE,5,10),即只对选择参数过程和地理投影过程执行并行.如果p_ncpu = 5，
则所有部分都使用5个cpu.
}
\examples{
maxent_auto(sp = system.file("extdata", "species/Phoebe sheareri.csv", package = "TBlabENM"),
evdir = system.file("extdata", "envar/asc", package = "TBlabENM"),
myenv = NULL,
evlist = 1:7,
factors = c("dr", "fao90"),
nbg = 10000,
args = maxent_args(),
fc = c("lph", "q", "lq"),
rm = 1:2,
r = 0.7,
vif = T,
vifth = 5,
opt = "auc.val.avg",
null_model = TRUE,
outdir = NULL)

}
